ui/requirements.txt
# aicli Desktop — backend dependencies only
# No ChromaDB, no sentence-transformers, no SQLite

# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# LLM clients (independent from CLI providers/)
anthropic>=0.40.0
openai>=1.50.0
google-genai>=1.0.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Config
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# PostgreSQL (optional — falls back to file store if DATABASE_URL not set)
psycopg2-binary>=2.9.9

# Utilities
pyyaml>=6.0
httpx>=0.27.0
